--- old/select/angular-select2.js
@@ -1,56 +1,0 @@
-define([
-  'angular',
-  'jquery',
-  'form/application/bootstrap',
-  'select2'
-], function(angular, $, app) {
-  'use strict';
-
-  var selectTwo = function($timeout) {
-
-    return {
-      restrict: 'A',
-      require: 'ngModel',
-      scope: {
-        config: '='
-      },
-      link: function(scope, element, attr, ngModel) {
-
-        var evalInParentScope = function(value) {
-          return scope.$parent.$eval(value);
-        };
-
-        var config = evalInParentScope(attr.config);
-
-        if (!angular.isDefined(config)) {
-          config = {};
-        }
-
-        $timeout(function() {
-          $(element).select2(config);
-        }, 0);
-
-        ngModel.$parsers.unshift(function(value) {
-          ngModel.$setValidity('schema', true);
-          return value;
-        });
-
-        // model -> DOM validation
-        ngModel.$formatters.unshift(function(value) {
-          // make it so that not selecting any of the choices is an error
-          if (typeof value === 'undefined') {
-            ngModel.$setValidity('schema', false);
-            return null;
-          }
-          // select2 stupidity: re-select the selected value
-          var selectedValue = $('option').filter(function() { return $(this).html() === value; }).val();
-          element.select2().select2('val', selectedValue);
-          return value;
-        });
-
-      }
-    };
-  };
-
-  return app.directive('selectTwo', selectTwo);
-});
