+++ src/modules/select/fixtures/angular-select2-fixture.html
