+++ src/modules/select/index.js
@@ -1,0 +1,8 @@
+module.exports = (function(){
+  angular.module('form-select', [
+    'schemaForm',
+    'form-decorators'
+  ])
+    .config(require('./form-selector'))
+    .directive('selectTwo', require('./angular-select2'));
+})();
